EVG-17347: Remove newlines from sample log files.
The jsonl had "\n" at the end of all the log lines from the python
script where I transformed it to jsonl. I'm guessing we're not
going to explicitly include the newline at the end of every log
line in our jsonl files so updated the test files to match.