EVG-17156: Basic infrastructure for fetching logs from S3.
This uses log_iterators copied over from Cedar, and takes advantage
of logs having an identical data format. There are a few changes
from the Cedar code:

1. The log iterators now filter chunks based on having any intersection
with their time range as opposed to being contained within their
time range. This is the biggest functional change from cedar's
iterators.
2. the LogChunkInfo struct has been modified to have a Key metho
rather than a Key value, to make it simpler to preserve the build
and test ids for later filter.

Also to note:
1. It looks like pail doesn't allow listing only one "level" of S3
keys, which means we pretty much always end up listing every
child of a build key. I think this should be fine from a
performance standpoint anyway, since we shouldn't have more than
1000ish subkeys anyway.
2. This code explicitly avoids using the MergeIterator across
every log file, because we would like to be able to load log
files in parallel. Instead it uses the BatchIterator. In principle
a parallel merge should be possible, but it is unnecessary
to preserve the existing behavior, which currently does not
correctly handle overlapping logs.