EVG-17155: Add functions to get global build log lines and all log lines. (#79)
* EVG-17155: Allow getting all log lines for a build.

This also fixes functionality for getting a specific test's logs, since we needed to get logs in the global logs that are before the next test but after the specified one

This also adds  a minimum and maximum time for time range to make
open ended time ranges easier. This was simpler than allowing
the ends to be "optional" since that substantially complicates
the business logic.