EVG-17149 store build (#80)
Add a method to storage to store build metadata in S3. 

I wrote it to rely on pulling S3 credentials out of the environment. This is already how we get our slack/evergreen credentials. Before we deploy this we will need to make sure that the credentials are in the application's environment or the application won't start.

Co-authored-by: John Daniels <john.daniels@mongodb.com>