EVG-17158: Add ability to view raw/html test logs via S3.
This adds an "s3" build parameter to view the logs for a test
that is stored in s3, and all the infrastructure necessary to
do that.

Interestingly, the current code for fetching log lines from mongo
doesn't handle errors, so when we translate our log iterator to
the existing interface, we silently ignore errors. We may want
to surface these better.