EVG-17160: Add featureswitch code for writing builds. (#98)
* EVG-17160: Add featureswitch code for writing builds.

This will allow us to set a percentage of builds being written to
S3. I opted to make this a hash of the build ID rather than random
for a few reasons:

1. When we use this for ramping up fetching from S3, it will ensure
that users won't alternate between code paths for fetching the same
build. This should avoid weird bugs for clients and aid in
debugging.
2. It will make debugging in general less based on chance. If you
put the same build name and id to the same configuration, you'll
hit the same code path.